Made the return type to be float32 explicitly out of the augmentation methods

+++ code/train_model.py
@@ -128,7 +128,7 @@ def speed_perturb(array, sr=16000):
     # "rate" resamples back to the original sample rate afterwards
     effects = [["speed", str(factor)], ["rate", str(sr)]]
     # sox_effects needs 2d tensor
-    waveform = torch.tensor(array).unsqueeze(0)
+    waveform = torch.tensor(array.astype(np.float32)).unsqueeze(0)
     waveform, _ = torchaudio.sox_effects.apply_effects_tensor(
             waveform, sr, effects
             )
@@ -139,7 +139,7 @@ def pitch_shift(array, sr = 16000):
     steps = random.randint(-2, 2)
     # sox pitch effect takes cents, semitones to cents by mult with 100
     effects = [["pitch", str(steps*100)], ["rate", str(sr)]]
-    waveform = torch.tensor(array).unsqueeze(0)
+    waveform = torch.tensor(array.astype(np.float32)).unsqueeze(0)
     waveform, _ = torchaudio.sox_effects.apply_effects_tensor(
             waveform, sr, effects
             )
